// src/shared/api/responses/bridge.response.ts
export type CreateBridgeResponse = {
	id: string;
	name: string;
	image: string;
};

export type GetAllBridgesResponse = {
	id: string;
	name: string;
	image: string;
}[];

export type GetOneBridgeResponse = {
	id: string;
	name: string;
	image: string;
};

export type UpdateBridgeResponse = {
	id: string;
	name: string;
	image: string;
};